refactor: :art: split all entities in archive but in unique type package
