+++ internal/domain/nft/nft.go
@@ -6,12 +6,6 @@ import (
 	"go.mongodb.org/mongo-driver/bson/primitive"
 )
 
-type NFTMetadata struct {
-	Title  string     `json:"title" bson:"title"`
-	Image  string     `json:"image" bson:"image"`
-	Expiry *time.Time `json:"expiry" bson:"expiry"`
-}
-
 type NFT struct {
 	ID        primitive.ObjectID `json:"id" bson:"id"` //Changed
 	Name      string             `json:"name" bson:"name"`
@@ -25,13 +19,5 @@ func NewNFT(name string, metadata NFTMetadata) *NFT {
 		Name:      name,
 		Metadata:  metadata,
 		CreatedAt: time.Now(),
-	}
-}
-
-func NewNFTMetadata(title, image string, expiry *time.Time) NFTMetadata {
-	return NFTMetadata{
-		Title:  title,
-		Image:  image,
-		Expiry: expiry,
 	}
 }
